code_medical/requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
medvae
pandas
numpy
matplotlib
Pillow
scikit-learn
